Rename kubeproxy-iptables package to kubeproxy_iptables
This PR also modifies the file and folder names to use
"_" instead of "-".
