+++ Makefile
@@ -40,8 +40,8 @@ reload-images: build images
 bin/submariner-engine: vendor/modules.txt main.go $(shell find pkg -not \( -path 'pkg/globalnet*' -o -path 'pkg/routeagent*' \))
 	${SCRIPTS_DIR}/compile.sh $@ main.go $(BUILD_ARGS)
 
-bin/submariner-route-agent: vendor/modules.txt $(shell find pkg/routeagent-driver)
-	${SCRIPTS_DIR}/compile.sh $@ ./pkg/routeagent-driver/main.go $(BUILD_ARGS)
+bin/submariner-route-agent: vendor/modules.txt $(shell find pkg/routeagent_driver)
+	${SCRIPTS_DIR}/compile.sh $@ ./pkg/routeagent_driver/main.go $(BUILD_ARGS)
 
 bin/submariner-globalnet: vendor/modules.txt $(shell find pkg/globalnet)
 	${SCRIPTS_DIR}/compile.sh $@ ./pkg/globalnet/main.go $(BUILD_ARGS)
